Fix: 가입신청서 생성 시 pickDropInfo requireItemList에서 제외 (#219)
* fix: 픽드랍 안내 필드 토글 제거
- 가입신청서 생성 시 픽드랍 안내 필드는 서버에 저장되지 않게 토글을 제거합니다.

* refactor: TicketType 리렌더 방지 memo 추가

* fix: SubmitButton에서 AlertBottomSheet 분리
- SubmitButton의 라이프사이클에서AlertBottomSheet를 제외합니다.
- 코드 리팩터링이 필요합니다.
