+++ src/components/Admin/EnrollmentForm/Stepper/SubmitButton.tsx
@@ -1,8 +1,6 @@
 import { getFieldStep } from "constants/step";
 
-import AlertBottomSheet from "components/common/BottomSheet/AlertBottomSheet";
 import { useAdminInfo } from "hooks/common/useAdminInfo";
-import { useOverlay } from "hooks/common/useOverlay";
 import { Adapter } from "libs/adapters";
 import { AdminFormToServerAdapter } from "libs/adapters/FormToServerAdapter";
 import { FieldErrors, FieldValues, useFormContext } from "react-hook-form";
@@ -15,34 +13,16 @@ import type { AdminEnrollmentInfoType } from "types/admin/enrollment.types";
 interface SubmitButtonProps {
   type?: "READ" | "CREATE" | "EDIT";
   onNextStep?: (formInfo: AdminEnrollmentInfoType) => void;
+  onOpenPopup: (field: string) => void;
 }
 
-const SubmitButton = ({ type, onNextStep }: SubmitButtonProps) => {
-  const { handleSubmit, setFocus } = useFormContext();
-  const overlay = useOverlay();
+const SubmitButton = ({ type, onNextStep, onOpenPopup }: SubmitButtonProps) => {
+  const { handleSubmit } = useFormContext();
 
   const setStep = useSetRecoilState(currentStepState);
   const { adminId, schoolId } = useAdminInfo();
 
   const text = type === "EDIT" ? "수정 완료" : "가입 신청서 등록";
-
-  const openPopup = (field: string) =>
-    overlay.open(({ isOpen, close }) => (
-      <AlertBottomSheet
-        isOpen={isOpen}
-        close={() => {
-          close();
-          setFocus(field);
-        }}
-        title="입력을 하지 않은 필수 항목이 있어요"
-        subtitle="유의사항에 동의하지 않으면 가입이 어려워요"
-        actionText="확인"
-        actionFn={() => {
-          close();
-          setFocus(field);
-        }}
-      />
-    ));
 
   const onSubmit = (data: FieldValues) => {
     const requestData = { ...data, adminId, schoolId };
@@ -57,7 +37,7 @@ const SubmitButton = ({ type, onNextStep }: SubmitButtonProps) => {
     const step = getFieldStep({ field: firstErrorField, enable: true });
     if (step !== undefined) {
       setStep(step);
-      openPopup(firstErrorField);
+      onOpenPopup(firstErrorField);
     }
   };
